requirements.txt
selenium
pytest
pytest-html
pytest-rerunfailures
pyyaml
allure-pytest
requests
pymysql
pytest-xdist
openpyxl